+++ src/scripts/types.d.ts
@@ -1,16 +1,16 @@
 // types.d.ts
 
 interface Point {
-    x: float;
-    y: float;
+    x: double;
+    y: double;
 } // Point
 
 interface Interval {
-    min: float;
-    max: float;
+    min: double;
+    max: double;
 } // Interval
 
 type int = number;
-type float = number;
+type double = number;
 type Vec3i = [int, int, int];
-type Vec3f = [float, float, float];
+type Vec3f = [double, double, double];
